+++ CHANGELOG.md
@@ -9,6 +9,9 @@
     - `seqkit fish`: fix range check. [#213](https://github.com/shenwei356/seqkit/issues/213)
     - `seqkit sample`: slightly faster.
     - `seqkit rmdup`: add flag for considering reverse complementary sequence [#215](https://github.com/shenwei356/seqkit/issues/215)
+    - `seqkit faidx`:
+        - supporting region with begin > end, i.e., return reverse complement sequence
+        - add new flag `-l/--region-file`:  file containing a list of regions
 - [SeqKit v0.16.1](https://github.com/shenwei356/seqkit/releases/tag/v0.16.1)
 [![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/seqkit/v0.16.1/total.svg)](https://github.com/shenwei356/seqkit/releases/tag/v0.16.1)
     - `seqkit shuffle --two-pass`: fix bug introduced in [#173](https://github.com/shenwei356/seqkit/issues/173) . [#209](https://github.com/shenwei356/seqkit/issues/209)